chore: :gear: 더미데이터 상수로 분리 #17
