+++ src/components/MapSidebar/index.tsx
@@ -1,3 +1,5 @@
+import ShopElement from 'components/MapSidebar/ShopElement';
+import { dummyShopList } from 'constants/dummy';
 import ActivePinIC from 'public/assets/ic_active_marker.svg';
 import InActivePinIC from 'public/assets/ic_basic_marker.svg';
 import EmptyStarIC from 'public/assets/ic_empty_star.svg';
@@ -6,69 +8,10 @@ import StarIC from 'public/assets/ic_star.svg';
 import { useState } from 'react';
 import styled, { css } from 'styled-components';
 
-import ShopElement, { ShopElementProps } from './ShopElement';
-
 interface StyledMapProps {
   isOpen?: boolean;
   isActive?: boolean;
 }
-
-export const dummyShopList: ShopElementProps[] = [
-  {
-    shopId: 2,
-    shopName: '수바코',
-    category: ['문구.팬시'],
-    image: '/assets/dummy/dummy-shop.png',
-    address: '서울 마포구 희우정로20길 66 상가1층2호',
-    close: '영업시간 12:00~17:00',
-    reviewCount: 42,
-  },
-  {
-    shopId: 3,
-    shopName: '프레젠트모먼트',
-    category: ['인테리어소품'],
-    image: '/assets/dummy/dummy-shop.png',
-    address: '서울 마포구 동교로 49-1 1층',
-    close: '영업시간 12:00~17:00',
-    reviewCount: 54,
-  },
-  {
-    shopId: 4,
-    shopName: '무드',
-    category: ['인테리어소품', '캔들'],
-    image: '/assets/dummy/dummy-shop.png',
-    address: '서울 마포구 포은로8길 16',
-    close: '영업시간 12:00~17:00',
-    reviewCount: 10,
-  },
-  {
-    shopId: 5,
-    shopName: '제로스페이스',
-    category: ['문구.팬시'],
-    image: '/assets/dummy/dummy-shop.png',
-    address: '서울 마포구 희우정로16길 32',
-    close: '영업시간 12:00~17:00',
-    reviewCount: 189,
-  },
-  {
-    shopId: 6,
-    shopName: '브라와',
-    category: ['문구.팬시'],
-    image: '/assets/dummy/dummy-shop.png',
-    address: '서울 마포구 희우정로10길 30 1층',
-    close: '영업시간 12:00~17:00',
-    reviewCount: 200,
-  },
-  {
-    shopId: 7,
-    shopName: '망원만물',
-    category: ['인테리어소품'],
-    image: '/assets/dummy/dummy-shop.png',
-    address: '서울 마포구 동교로 35 왼쪽가게',
-    close: '영업시간 12:00~17:00',
-    reviewCount: 200,
-  },
-];
 
 type OptionLabel = '인기 순' | '내가 저장한';
 interface OptionInfo {
@@ -156,7 +99,6 @@ const OptionList = styled.ul`
   flex-direction: column;
   height: 100%;
   border-right: 0.1rem solid ${({ theme }) => theme.colors.gray2};
-  margin-left: -1px;
 `;
 
 const Option = styled.li<StyledMapProps>`
@@ -167,7 +109,7 @@ const Option = styled.li<StyledMapProps>`
   gap: 0.6rem;
 
   width: 100%;
-  height: 7.8rem;
+  height: 7.65rem;
 
   ${(props) =>
     props.isActive
